Migrate feedback management to single-page system
